Splitting out commit commands
