+++ Commands/svnCommitHistoryCommand.py
@@ -1,0 +1,34 @@
+import sublime
+import sublime_plugin
+
+from ..Core.Controller import *
+
+class svnCommitHistoryCommand(sublime_plugin.TextCommand, svnController):
+    def run(self, edit):
+        self.svnDir = self.get_svn_dir()
+        if len(self.svnDir) == 0:
+            return;
+
+        self.svnDir = self.get_scoped_path(svn_settings().get('SVN.commit_scope', 'file'))
+
+        self.fileList = list(svn_settings().get('SVN.history', []))
+        self.fileList.insert(min(len(self.fileList), 1), 'New Log')
+
+        sublime.active_window().show_quick_panel(self.fileList, self.on_ticket)
+
+    def on_ticket(self, index):
+        try:
+            message = self.fileList[index]
+
+            if index == -1:
+                pass
+            elif message == 'New Log':
+                sublime.active_window().run_command('svn_commit')
+            else:
+                self.do_commit(message)
+
+        except ValueError:
+            pass
+
+    def is_enabled(self):
+        return len(str(self.get_svn_dir())) != 0
