+++ SVNTools.py
@@ -14,6 +14,11 @@ from .Commands.svnUpdateRepoCommand import *
 from .Commands.svnDiscardChangesCommand import *
 from .Commands.svnShowChangesCommand import *
 
+from .Commands.svnCommitCommand import *
+from .Commands.svnCommitLastCommand import *
+from .Commands.svnCommitBlankCommand import *
+from .Commands.svnCommitHistoryCommand import *
+
 
 def plugin_loaded():
     sublime.avibeSVNToolsTicketNo = ""
@@ -25,159 +30,11 @@ def plugin_loaded():
     print("loaded")
     print("==============")
 
-def svn_settings():
-    return sublime.load_settings( 'Preferences.sublime-settings' )
-
-def svn_set_status_items(self, view):
-
-    if svn_settings().get('SVN.show_status_bar_info', 1) == 1:
-        self.svnDir = self.get_svn_dir()
-        if len(self.svnDir) == 0:
-            view.set_status('AAAsvnTool', 'SVN:' + u'\u2718')
-            view.erase_status('AABsvnTool')
-            view.erase_status('AACsvnTool')
-        else:
-            view.set_status('AAAsvnTool', 'SVN:' + u'\u2714')
-            view.set_status('AABsvnTool', 'Scope: ' + str(svn_settings().get('SVN.commit_scope', 'repo')))
-            if svn_settings().get('SVN.show_diff_in_status_bar', 0) == 1:
-                self.svnDir = self.get_scoped_path('file')
-                procText = self.run_svn_command([ "svn", "status", self.svnDir])
-
-                if len(procText):
-                    procText = procText.split( '\n' )[0].strip( )
-                    procText = procText.split( )[0]
-
-                    if procText == 'M':
-                        view.set_status('AACsvnTool', 'diff:' + u'\u2260' )
-                    else:
-                        view.set_status('AACsvnTool', 'diff:' + u'\u003D' )
-                else:
-                    view.set_status('AACsvnTool', 'diff:' + u'\u003D' )
-
-
-            else:
-                view.erase_status('AACsvnTool')
-    else:
-        view.erase_status('AAAsvnTool')
-        view.erase_status('AABsvnTool')
-        view.erase_status('AACsvnTool')
-
-
-
-
-
 class ChangeLog(sublime_plugin.TextCommand, svnController):
     def run(self, edit):
         procText = self.run_svn_command([ "svn", "log", "-v", self.svnDir]);
 
         show_output_panel(procText)
-
-
-
-class svnCommitCommand(sublime_plugin.TextCommand, svnController):
-    def run(self, edit):
-        self.svnDir = self.get_svn_dir()
-        if len(self.svnDir) == 0:
-            return;
-
-        self.svnDir = self.get_scoped_path(svn_settings().get('SVN.commit_scope', 'repo'))
-
-        sublime.active_window().run_command("save")
-        sublime.active_window().show_input_panel("Ticket number:", sublime.avibeSVNToolsTicketNo, self.on_ticket, None, None)
-        pass
-
-    def on_ticket(self, text):
-        try:
-            sublime.avibeSVNToolsTicketNo = text
-
-            sublime.active_window().show_input_panel("Comment:", sublime.avibeSVNToolsThisComment, self.on_comment, None, None)
-        except ValueError:
-            pass
-
-    def on_comment(self, text):
-        sublime.avibeSVNToolsThisComment = text
-        sublime.avibeSVNToolsLastComment = sublime.avibeSVNToolsTicketNo
-
-        if len( sublime.avibeSVNToolsLastComment ):
-            sublime.avibeSVNToolsLastComment = "#" + sublime.avibeSVNToolsLastComment + ": "
-
-        sublime.avibeSVNToolsLastComment = sublime.avibeSVNToolsLastComment + sublime.avibeSVNToolsThisComment
-
-        self.do_commit(sublime.avibeSVNToolsLastComment)
-
-    def is_enabled(self):
-        return len(str(self.get_svn_dir())) != 0
-
-class svnCommitLastCommand(sublime_plugin.TextCommand, svnController):
-    def run(self, edit):
-        self.svnDir = self.get_svn_dir()
-        if len(self.svnDir) == 0:
-            return;
-
-        if( 0 == len( sublime.avibeSVNToolsLastComment )):
-            sublime.status_message( "Commit with comment (CTRL-ALT-B twice) to use this shortcut." );
-            return
-
-        self.svnDir = self.get_scoped_path(svn_settings().get('SVN.commit_scope', 'repo'))
-
-        sublime.active_window().run_command("save")
-
-        self.do_commit(sublime.avibeSVNToolsLastComment)
-
-    def is_enabled(self):
-        return len(str(self.get_svn_dir())) != 0
-
-class svnCommitBlankCommand(sublime_plugin.TextCommand, svnController):
-    def run(self, edit):
-        self.svnDir = self.get_svn_dir()
-        if len(self.svnDir) == 0:
-            return;
-
-        self.svnDir = self.get_scoped_path(svn_settings().get('SVN.commit_scope', 'repo'))
-
-        sublime.active_window().run_command("save")
-
-        self.do_commit("")
-
-    def is_enabled(self):
-        return len(str(self.get_svn_dir())) != 0
-
-class svnCommitHistoryCommand(sublime_plugin.TextCommand, svnController):
-    def run(self, edit):
-        self.svnDir = self.get_svn_dir()
-        if len(self.svnDir) == 0:
-            return;
-
-        self.svnDir = self.get_scoped_path(svn_settings().get('SVN.commit_scope', 'repo'))
-
-        self.fileList = list(svn_settings().get('SVN.history', []))
-        self.fileList.insert(min(len(self.fileList), 1), 'New Log')
-
-        sublime.active_window().show_quick_panel(self.fileList, self.on_ticket)
-
-    def on_ticket(self, index):
-        try:
-            message = self.fileList[index]
-
-            if index == -1:
-                pass
-            elif message == 'New Log':
-                sublime.active_window().run_command('svn_commit')
-            else:
-                self.do_commit(message)
-
-        except ValueError:
-            pass
-
-    def is_enabled(self):
-        return len(str(self.get_svn_dir())) != 0
-
-
-
-
-
-
-
 
 class svnLogParser(svnController):
     def run(self):
